SelArray: switch to a base+offset representation
This is necessary to introduce more equations about
merge (e.g. associativity) and a ghost version of split.

In fact, I split the implementation into two parts:
managing a sequence of permissions, where the array
is a sequence of refs as before, and managing the
array proper.

In this new representation, the array has a
base array into which to index, as the
base allocation unit, and a from and to
indices. I wish "to" were ghost (thus more
accurately modeling C pointers), but this
would need introducing ghost effects for
SteelSel.